Update main buttons to fix edge click cases
Redesigned the implementation of main buttons and added some customs attributes to CopyHelper component.
Refactored Copy.tsx just a little to not get confused between CopyIcon, CopyButton, CopyHelper. Main Copy button has its own component CopyBtn now.
Twitter button all the same just made the parent to be a link as well and added the 'href' attribute to it instead of to the elements "inside" the button.
minWidth and minHeight changes to CopyHelper references for layout consistency.